Restructure project layout and Docker assets
Move core source files into Config, Data, Game and Server namespaces and folders, add GlobalUsings.cs to simplify imports, and update affected namespaces. Relocate Dockerfiles and entrypoint scripts into docker/ and update docker-compose and Dockerfile COPY paths accordingly. Minor fixes: change GameManager.teamCapacity type from uint to ushort, add .dockerignore entry for docs, and add .claude local settings file.

+++ Components/Blaze/GameManager.cs
@@ -173,7 +173,7 @@ internal class GameManager : GameManagerBase.Server
             // }
 
             var gameMode = serverGame.ReplicatedGameData.mGameAttribs.TryGetValue("OSDK_gameMode", out var mode) ? mode : "1";
-            var teamCapacity = gameMode == "3" ? (uint)6 : (uint)1;
+            var teamCapacity = gameMode == "3" ? (ushort)6 : (ushort)1;
 
             lobbies.Add(new GameBrowserMatchData
             {
